Handle race condition where StateFile would see an empty file
case CPANEL-15077

In StateFile.pm we generate an new taskqueue file in the
event it has been wiped out.  Unfortunately the check to
see if the file was zero bytes was done without a lock
file which resulted in a false positive when another
process was writing the file.  This change re-checks
after obtaing a lock to ensure the file really is zero
and not just being updated.