Avoid giving up a lock when we are forking
case CPANEL-17403

The TaskQueue would give up the lock on the state file in order to
process as task.  Since cPanel processes most of our tasks with the
cPanel::TaskQueue::ChildProcessor task in the background this added a
write state, unlock, relock, re-read from disk cycle to every task
being processed.  Since we return right after queueprocd fork()s there
is little concern about holding the lock open.

This should also reduce the potential for unknown race conditions as we
no longer give up the lock in the middle of processing a task which had
caused unexpected results in the past.