Avoid duplicate tasks when moving out of deferred queue
Case CPANEL-21194: Normally when we queue a task we process overrides
and check for duplicates.  However, we fail to do so when moving items
from the deferred queue to the processing queue, potentially resulting
in duplicate items.  Perform these tasks when moving items to avoid
these duplicates.  Since the functions used operate on the waiting
queue, ensure that we manipulate the waiting queue directly to avoid
queuing duplicate items if the deferred queue has duplicates.