Don't confuse callers by returning a failure if plugin already included
Case CPANEL-22183: It's common practice to do:

    cPanel::TaskQueue::PluginManager::load_plugin_by_name(...) or ...

And it can't reasonably be considered that skipping setup because
it is already setup is a *failure* condition.  Ergo, we make it into
a success condition.  This will fix a number of existing callers making
this assumption incorrectly.

Changelog: Make plugin include return true when plugins are already
        included.