Bump the file size on change
Fix #34

Recent change in blead exposed a bug in `cPanel::StateFile::_resynch`.

When using multiple cPanel::TaskQueue objects, communication between
both objects and `resync` is based on timestamp and filesize to force
reloading the cache.

This change is introducing an additional field to the serialized data to
bump the file size (and eventually rotate after reaching a certain size)
on updates where there would be a no-size change.

Refs: https://github.com/Perl/perl5/issues/19532